+++ CHANGELOG.md
@@ -2,7 +2,16 @@
 
 ## Unreleased
 
-No changes yet!
+### Build changes
+
+- Use [setuptools-scm](https://pypi.org/project/setuptools-scm/) for version
+  management. This is a breaking change, as it removes the
+  `acc_app_optimisation.__version__` constant. Please use
+  [`importlib.metadata`](https://docs.python.org/3/library/importlib.metadata.html)
+  (backported as
+  [`importlib_metadata`](https://importlib_metadata.readthedocs.io/) before
+  Python 3.8) to query the app's version programmatically.
+
 
 ## v0.3.4